fix: update deprecated APIs and remove code quality warnings
Fixes #2058

Update deprecated API calls and remove unnecessary code patterns across the codebase.

**Deprecated API Updates:**

1. **Modifier.menuAnchor()** (MenuCustomizationScreen.kt)
   - Use new ExposedDropdownMenuAnchorType parameter
   - Specify PrimaryNotEditable type and enabled state

2. **ImageAnalysis.setTargetResolution()** (CameraForegroundService.kt)
   - Migrate to ResolutionSelector API
   - Use ResolutionFilter for more flexible resolution selection

3. **Drawable.getOpacity()** (PointScanBlockDrawable.kt)
   - Add @Deprecated annotation to override

**Code Quality Improvements:**

4. **Remove unnecessary safe calls** (4 occurrences):
   - SwitchActionField.kt: savedStateHandle is non-null in context
   - LinearGesturePerformer.kt: accessibilityService is non-null constructor parameter
   - MainMenuStructure.kt: getPauseManager() returns non-null
   - NodeExaminer.kt: getBounds() returns non-null Rect (3 occurrences)

5. **Remove redundant conversions** (6 occurrences):
   - ItemScanSettingsView.kt: String.toString() → String
   - PointScanSettingsView.kt: String.toString() → String
   - PinchGesturePerformer.kt: Float.toFloat() → Float (2 occurrences)
   - SelectionHandler.kt: String.toString() → String (2 occurrences)

**Files Changed (11 total):**
- MenuCustomizationScreen.kt
- CameraForegroundService.kt
- PointScanBlockDrawable.kt
- SwitchActionField.kt
- ItemScanSettingsView.kt
- PointScanSettingsView.kt
- LinearGesturePerformer.kt
- PinchGesturePerformer.kt
- MainMenuStructure.kt
- NodeExaminer.kt
- SelectionHandler.kt

**Benefits:**
- Removes 20+ compiler warnings
- Uses modern, supported APIs
- Cleaner, more efficient code
- Better IDE experience
- Future-proof against API removals

+++ app/src/main/java/com/enaboapps/switchify/service/gestures/PinchGesturePerformer.kt
@@ -61,8 +61,8 @@ object PinchGesturePerformer {
 
         // Show visual feedback
         pinchVisual?.start(
-            centerPoint.x.toFloat(),
-            centerPoint.y.toFloat(),
+            centerPoint.x,
+            centerPoint.y,
             visualCircleSize,
             GestureData.PINCH_DURATION,
             type == GestureType.PINCH_OUT
